requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==4.2.13
django-cors-headers==4.3.1
django-mssql-backend==2.8.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-extensions==3.2.3
mssql-django==1.5
mypy==1.6.1
mypy-extensions==1.0.0
celery==5.4.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
PyJWT==2.8.0
PyMySQL==1.1.1
pyodbc==5.1.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
cryptography==42.0.7
channels==4.1.0
daphne==4.1.2
redis==5.0.4
channels-redis==4.2.0
python-dateutil==2.9.0.post0
python-crontab==3.1.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
Werkzeug==3.0.3
requests==2.32.3